crear interfaces separadas
