+++ web/src/interfaces/category.ts
@@ -1,0 +1,5 @@
+export interface CategoryProps {
+  to: string;
+  label: string;
+  img: string;
+}
