+++ web/src/interfaces/checkout.ts
@@ -1,0 +1,12 @@
+export interface CheckoutProps {
+  total: number;
+  onSubmit: (
+    formData: { 
+      name: string; 
+      paymentMethod: string; 
+      address: string; 
+      phone: string 
+    },
+  ) => void;
+  goBack: () => void;
+}
