+++ web/src/interfaces/product.ts
@@ -1,0 +1,13 @@
+export interface Product {
+  id: number;
+  name: string;
+  price: number;
+  description: string;
+  brand: string;
+  weight: string;
+  image: string;
+}
+
+export interface ProductProps {
+  product: Product;
+}
